Add LTE function pins and data send code.
Added PA5 for LTE_ON and PA6 for LTE_RST. ON is pulled down, RST is pulled up.

For data send code, added AT commands sent over UART to interact with LTE module and provide a connection to the server. Initial setup is done, but untested. SD card read and message publishing still mist be done.